change owner_id type from int to long
it's set as int8 in the artworks db schema, and is being assumed as long in all tests and apis, which is why both tests failed.

+++ code/VR Art Gallery/Assets/Scripts/Cloud/Models/ArtworkData.cs
@@ -16,7 +16,7 @@ public class ArtworkData : BaseModel
     public string title { get; set; }
 
     [Column("owner_id")]
-    public int owner_id { get; set; }
+    public long owner_id { get; set; }
   
     // Image storage
     [Column("image_url")]
